mongoDB 연결 app.py에 통합
