+++ app/db.py
@@ -1,6 +1,7 @@
 from motor.motor_asyncio import AsyncIOMotorClient
 
-mongoDB_url: str = "mongodb://127.0.0.1:27017"
+# DB 정보
+mongoDB_url: str = "mongodb://192.168.0.141:27017"
 database_name: str = "ohmyuchu"
 
 # DB 리소스 관리 객체
